Move API tests into their own file
